[add] create `pci::config_register` module
